% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_project.R
\name{initialize_project}
\alias{initialize_project}
\title{Initialize project}
\usage{
initialize_project(project_name, directory_parent = getwd())
}
\arguments{
\item{project_name}{directory name for project}

\item{directory_parent}{parent directory for project}
}
\description{
Initialize project
}
